Refactor Repository and Service, clean authentication and authorization

+++ backend/src/main/java/com/socialapp/backend/like/LikeRepositoryImpl.java
@@ -2,53 +2,62 @@ package com.socialapp.backend.like;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcOperations;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.stereotype.Repository;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 @Repository
 @RequiredArgsConstructor
 public class LikeRepositoryImpl implements LikeRepository {
-    private final JdbcTemplate jdbcTemplate;
+    private final NamedParameterJdbcTemplate jdbcTemplate;
 
     @Override
     public boolean isUserLikedPost(Long id, Long userId) {
-        String sql = "SELECT COUNT(*) FROM `like_posts` WHERE post_id = ? AND user_id = ?";
-        Object[] params = new Object[]{id, userId};
+        String sql = "SELECT COUNT(*) FROM `like_posts` WHERE post_id = :id AND user_id = :userId";
 
-        Integer num = jdbcTemplate.queryForObject(sql, Integer.class, params);
+        Map<String, Object> params = new HashMap<>();
+        params.put("id", id);
+        params.put("userId", userId);
 
-        return num != 0;
+        Integer num = jdbcTemplate.queryForObject(sql,  params, Integer.class);
+
+        return num != null && num > 0;
     }
 
     @Override
-    public void insert(Like like) {
+    public int insert(Like like) {
         String sql = "INSERT INTO like_posts(`post_id`, `user_id`) VALUES (:postId, :userId)";
 
         SqlParameterSource params = new BeanPropertySqlParameterSource(like);
-        NamedParameterJdbcOperations template = new NamedParameterJdbcTemplate(jdbcTemplate);
 
-        template.update(sql, params);
+        return jdbcTemplate.update(sql, params);
     }
 
     @Override
-    public void delete(Long id, Long userId) {
-        String sql = "DELETE FROM `like_posts` WHERE post_id = ? AND user_id = ?";
-        this.jdbcTemplate.update(sql, id, userId);
+    public int delete(Long id, Long userId) {
+        String sql = "DELETE FROM `like_posts` WHERE post_id = :id AND user_id = :userId";
+
+        Map<String, Object> params = new HashMap<>();
+        params.put("id", id);
+        params.put("userId", userId);
+
+        return this.jdbcTemplate.update(sql, params);
     }
 
     @Override
     public Optional<List<Like>> findAllByPostId(Long id) {
-        String sql = "SELECT * FROM like_posts li WHERE li.post_id = ?";
-        Object[] params = new Object[]{id};
+        String sql = "SELECT * FROM like_posts li WHERE li.post_id = :id";
 
-        List<Like> likes = jdbcTemplate.query(sql, BeanPropertyRowMapper.newInstance(Like.class), params);
+        Map<String, Object> params = new HashMap<>();
+        params.put("id", id);
+
+        List<Like> likes = jdbcTemplate.query(sql, params, BeanPropertyRowMapper.newInstance(Like.class));
         return Optional.of(likes);
     }
 }
